some refactoring and extract the breaking values
