+++ packages/data-context/src/sources/migration/codegen.ts
@@ -1,4 +1,3 @@
-import chokidar from 'chokidar'
 import fs from 'fs-extra'
 import path from 'path'
 import globby from 'globby'
@@ -11,7 +10,6 @@ import prettier from 'prettier'
 import type { MigrationFile } from '../MigrationDataSource'
 import { supportFileRegexps } from './regexps'
 import Debug from 'debug'
-import { toPosix } from '../../util'
 
 const debug = Debug('cypress:data-context:sources:migration:codegen')
 
@@ -57,95 +55,6 @@ export interface CreateConfigOptions {
 
 export async function createConfigString (cfg: OldCypressConfig, options: CreateConfigOptions) {
   return createCypressConfig(reduceConfig(cfg), await getPluginRelativePath(cfg, options.projectRoot), options)
-}
-
-interface FileToBeMigratedManually {
-  relative: string
-  moved: boolean
-}
-
-export interface ComponentTestingMigrationStatus {
-  files: Map<string, FileToBeMigratedManually>
-  completed: boolean
-}
-
-export async function initComponentTestingMigration (
-  projectRoot: string,
-  componentFolder: string,
-  testFiles: string[],
-  onFileMoved: (status: ComponentTestingMigrationStatus) => void,
-): Promise<{
-  status: ComponentTestingMigrationStatus
-  watcher: chokidar.FSWatcher | null
-}> {
-  debug('initComponentTestingMigration %O', { projectRoot, componentFolder, testFiles })
-  const watchPaths = testFiles.map((glob) => {
-    return `${componentFolder}/${glob}`
-  })
-
-  const watcher = chokidar.watch(
-    watchPaths, {
-      cwd: projectRoot,
-    },
-  )
-
-  debug('watchPaths %o', watchPaths)
-
-  let filesToBeMoved: Map<string, FileToBeMigratedManually> = (await globby(watchPaths, {
-    cwd: projectRoot,
-  })).reduce<Map<string, FileToBeMigratedManually>>((acc, relative) => {
-    acc.set(relative, { relative, moved: false })
-
-    return acc
-  }, new Map())
-
-  debug('files to be moved manually %o', filesToBeMoved)
-  if (filesToBeMoved.size === 0) {
-    // this should not happen as the step should be hidden in this case
-    // but files can have been moved manually before clicking next
-    return {
-      status: {
-        files: filesToBeMoved,
-        completed: true,
-      },
-      watcher: null,
-    }
-  }
-
-  watcher.on('unlink', (unlinkedPath) => {
-    const posixUnlinkedPath = toPosix(unlinkedPath)
-    const file = filesToBeMoved.get(posixUnlinkedPath)
-
-    if (!file) {
-      throw Error(`Watcher incorrectly triggered ${posixUnlinkedPath}
-      while watching ${Array.from(filesToBeMoved.keys()).join(', ')}
-      projectRoot: ${projectRoot}`)
-    }
-
-    file.moved = true
-
-    const completed = Array.from(filesToBeMoved.values()).every((value) => value.moved === true)
-
-    onFileMoved({
-      files: filesToBeMoved,
-      completed,
-    })
-  })
-
-  return new Promise((resolve, reject) => {
-    watcher.on('ready', () => {
-      debug('watcher ready')
-      resolve({
-        status: {
-          files: filesToBeMoved,
-          completed: false,
-        },
-        watcher,
-      })
-    }).on('error', (err) => {
-      reject(err)
-    })
-  })
 }
 
 async function getPluginRelativePath (cfg: OldCypressConfig, projectRoot: string): Promise<string> {
@@ -329,13 +238,13 @@ export function renameSupportFilePath (relative: string) {
 export function reduceConfig (cfg: OldCypressConfig): ConfigOptions {
   const excludedFields = ['pluginsFile', '$schema']
 
-  return Object.entries(cfg).reduce((acc, [key, val]) => {
+  const reducedConfig = Object.entries(cfg).reduce((acc, [key, val]) => {
     if (excludedFields.includes(key)) {
       return acc
     }
 
     if (key === 'e2e' || key === 'component') {
-      const value = val as Cypress.ResolvedConfigOptions
+      const value = val as OldCypressConfig
 
       if (!value) {
         return acc
@@ -413,6 +322,10 @@ export function reduceConfig (cfg: OldCypressConfig): ConfigOptions {
 
     return { ...acc, global: { ...acc.global, [key]: val } }
   }, { global: {}, e2e: {}, component: {} })
+
+  debug('reduceConfig: %O', reducedConfig)
+
+  return reducedConfig
 }
 
 function getSpecPattern (cfg: OldCypressConfig, testType: TestingType) {
