+++ packages/data-context/src/sources/migration/index.ts
@@ -4,5 +4,6 @@
 export * from './autoRename'
 export * from './codegen'
 export * from './format'
+export * from './manualRenameWatcher'
 export * from './regexps'
 export * from './shouldShowSteps'
